requirements.txt

flask
slack-sdk
python-dotenv
requests
